fix bugs on back and frontend changes

+++ TaxiApp/UsersService/UsersService.cs
@@ -48,19 +48,19 @@ namespace UsersService
 
         public async Task<bool> addNewUser(User user)
         {
-            var userDictionary = await StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var userDictionary = await StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
 
             try
             {
                 using (var transaction = StateManager.CreateTransaction())
                 {
-                    if (!await CheckIfUserAlreadyExists(user)) 
+                    if (!await CheckIfUserAlreadyExists(user))
                     {
 
-                        await userDictionary.AddAsync(transaction, user.Email, user); // dodaj ga prvo u reliable 
+                        await userDictionary.AddAsync(transaction, user.Id, user); // dodaj ga prvo u reliable 
 
                         //insert image of user in blob
-                        CloudBlockBlob blob = await dataRepo.GetBlockBlobReference("users", $"image_{user.Username}");
+                        CloudBlockBlob blob = await dataRepo.GetBlockBlobReference("users", $"image_{user.Id}");
                         blob.Properties.ContentType = user.ImageFile.ContentType;
                         await blob.UploadFromByteArrayAsync(user.ImageFile.FileContent, 0, user.ImageFile.FileContent.Length);
                         string imageUrl = blob.Uri.AbsoluteUri;
@@ -76,7 +76,7 @@ namespace UsersService
                     }
                     return false;
                 }
-               
+
             }
             catch (Exception ex)
             {
@@ -86,14 +86,25 @@ namespace UsersService
 
         private async Task<bool> CheckIfUserAlreadyExists(User user)
         {
-            var users = await StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var users = await StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
             try
             {
-                using (var transaction = StateManager.CreateTransaction())
+                using (var tx = StateManager.CreateTransaction())
                 {
-                    var result = await users.TryGetValueAsync(transaction, user.Email);
-                    return result.HasValue;
+                    var enumerable = await users.CreateEnumerableAsync(tx);
+
+                    using (var enumerator = enumerable.GetAsyncEnumerator())
+                    {
+                        while (await enumerator.MoveNextAsync(default(CancellationToken)))
+                        {
+                            if (enumerator.Current.Value.Email == user.Email || enumerator.Current.Value.Id == user.Id || enumerator.Current.Value.Username == user.Username)
+                            {
+                                return true;
+                            }
+                        }
+                    }
                 }
+                return false;
             }
             catch (Exception)
             {
@@ -105,7 +116,7 @@ namespace UsersService
 
         private async Task LoadUsers()
         {
-            var userDictionary = await StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var userDictionary = await StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
             try
             {
                 using (var transaction = StateManager.CreateTransaction())
@@ -117,8 +128,8 @@ namespace UsersService
                         foreach (var user in users)
                         {
                             byte[] image = await dataRepo.DownloadImage(dataRepo, user, "users");
-                            await userDictionary.AddAsync(transaction, user.Email, UserMapper.MapUserEntityToUser(user,image)); 
-                        } 
+                            await userDictionary.AddAsync(transaction, user.Id, UserMapper.MapUserEntityToUser(user, image));
+                        }
                     }
 
                     await transaction.CommitAsync();
@@ -154,7 +165,7 @@ namespace UsersService
             //       or remove this RunAsync override if it's not needed in your service.
 
             var myDictionary = await this.StateManager.GetOrAddAsync<IReliableDictionary<string, long>>("myDictionary");
-            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
             // ovde ce biti load users 
             await LoadUsers();
             while (true)
@@ -178,25 +189,32 @@ namespace UsersService
                 await Task.Delay(TimeSpan.FromSeconds(1), cancellationToken);
             }
         }
+        public async Task<LogedUserDTO> loginUser(LoginUserDTO loginUserDTO)
+        {
+            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
 
-        public async Task<FullUserDTO> loginUser(LoginUserDTO loginUserDTO)
-          {
-            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
             using (var tx = this.StateManager.CreateTransaction())
             {
-                ConditionalValue<User> result = await users.TryGetValueAsync(tx, loginUserDTO.Email);
-                if (result.HasValue && result.Value.Email == loginUserDTO.Email && result.Value.Password == loginUserDTO.Password)
-                {
-                   return new FullUserDTO(result.Value.Address,result.Value.AverageRating,result.Value.SumOfRatings,result.Value.NumOfRatings,result.Value.Birthday,result.Value.Email,result.Value.IsVerified,result.Value.IsBlocked,result.Value.FirstName,result.Value.LastName,result.Value.Username,result.Value.TypeOfUser,result.Value.ImageFile,result.Value.Password,result.Value.Status);
-                }
-                else return new FullUserDTO();
+                var enumerable = await users.CreateEnumerableAsync(tx);
 
+                using (var enumerator = enumerable.GetAsyncEnumerator())
+                {
+                    while (await enumerator.MoveNextAsync(default(CancellationToken)))
+                    {
+                        if (enumerator.Current.Value.Email == loginUserDTO.Email && enumerator.Current.Value.Password == loginUserDTO.Password)
+                        {
+                            return new LogedUserDTO(enumerator.Current.Value.Id,enumerator.Current.Value.TypeOfUser);
+                        }
+                    }
+                }
             }
+            return new LogedUserDTO();
         }
+
 
         public async Task<List<FullUserDTO>> listUsers()
         {
-            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
 
             List<FullUserDTO> userList = new List<FullUserDTO>();
 
@@ -219,40 +237,40 @@ namespace UsersService
 
         public async Task<List<DriverViewDTO>> listDrivers()
         {
-            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
             List<DriverViewDTO> drivers = new List<DriverViewDTO>();
             using (var tx = this.StateManager.CreateTransaction())
             {
                 var enumerable = await users.CreateEnumerableAsync(tx);
                 using (var enumerator = enumerable.GetAsyncEnumerator())
                 {
-                    while(await enumerator.MoveNextAsync(default (CancellationToken)))
+                    while (await enumerator.MoveNextAsync(default(CancellationToken)))
                     {
-                        if(enumerator.Current.Value.TypeOfUser == UserRoles.Roles.Driver)
+                        if (enumerator.Current.Value.TypeOfUser == UserRoles.Roles.Driver)
                         {
-                            drivers.Add(new DriverViewDTO(enumerator.Current.Value.Email, enumerator.Current.Value.FirstName, enumerator.Current.Value.LastName, enumerator.Current.Value.Username, enumerator.Current.Value.IsBlocked,enumerator.Current.Value.AverageRating));
+                            drivers.Add(new DriverViewDTO(enumerator.Current.Value.Email, enumerator.Current.Value.FirstName, enumerator.Current.Value.LastName, enumerator.Current.Value.Username, enumerator.Current.Value.IsBlocked, enumerator.Current.Value.AverageRating,enumerator.Current.Value.Id));
                         }
                     }
                 }
             }
 
-            return drivers; 
+            return drivers;
 
         }
 
-        public async Task<bool> changeDriverStatus(string email, bool status)
+        public async Task<bool> changeDriverStatus(Guid id, bool status)
         {
-            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
             using (var tx = this.StateManager.CreateTransaction())
             {
-                ConditionalValue<User> result = await users.TryGetValueAsync(tx,email);
+                ConditionalValue<User> result = await users.TryGetValueAsync(tx, id);
                 if (result.HasValue)
                 {
                     User user = result.Value;
                     user.IsBlocked = status;
-                    await users.SetAsync(tx, email, user);
+                    await users.SetAsync(tx, id, user);
 
-                    await dataRepo.UpdateEntity(email, status);
+                    await dataRepo.UpdateEntity(id, status);
 
                     await tx.CommitAsync();
 
@@ -267,56 +285,69 @@ namespace UsersService
 
         public async Task<FullUserDTO> changeUserFields(UserForUpdateOverNetwork user)
         {
-            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<string, User>>("UserEntities");
+            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
             using (var tx = this.StateManager.CreateTransaction())
             {
-                ConditionalValue<User> result = await users.TryGetValueAsync(tx, user.PreviousEmail);
+                ConditionalValue<User> result = await users.TryGetValueAsync(tx, user.Id);
                 if (result.HasValue)
                 {
                     User userFromReliable = result.Value;
-                    string previousUsername = userFromReliable.Username;
 
-                    if (!string.IsNullOrEmpty(user.Email))  userFromReliable.Email = user.Email;
+                    if (!string.IsNullOrEmpty(user.Email)) userFromReliable.Email = user.Email;
 
-                    if (!string.IsNullOrEmpty(user.FirstName)) userFromReliable.FirstName = user.FirstName;    
-                    
-                    if(!string.IsNullOrEmpty(user.LastName)) userFromReliable.LastName = user.LastName;
-                    
-                    if(!string.IsNullOrEmpty(user.Address)) userFromReliable.Address = user.Address;
+                    if (!string.IsNullOrEmpty(user.FirstName)) userFromReliable.FirstName = user.FirstName;
 
-                    if(user.Birthday != DateTime.MinValue) userFromReliable.Birthday = user.Birthday;
+                    if (!string.IsNullOrEmpty(user.LastName)) userFromReliable.LastName = user.LastName;
+
+                    if (!string.IsNullOrEmpty(user.Address)) userFromReliable.Address = user.Address;
+
+                    if (user.Birthday != DateTime.MinValue) userFromReliable.Birthday = user.Birthday;
 
                     if (!string.IsNullOrEmpty(user.Password)) userFromReliable.Password = user.Password;
 
                     if (!string.IsNullOrEmpty(user.Username)) userFromReliable.Username = user.Username;
 
-                    if (user.ImageFile.FileContent.Length > 0) userFromReliable.ImageFile = user.ImageFile;
+                    if (user.ImageFile != null && user.ImageFile.FileContent != null && user.ImageFile.FileContent.Length > 0) userFromReliable.ImageFile = user.ImageFile;
 
-                        await users.TryRemoveAsync(tx,user.PreviousEmail); // ukloni ovog proslog 
+                    await users.TryRemoveAsync(tx, user.Id); // ukloni ovog proslog 
 
-                        await users.AddAsync(tx, userFromReliable.Email, userFromReliable); // dodaj ga prvo u reliable 
+                    await users.AddAsync(tx, userFromReliable.Id, userFromReliable); // dodaj ga prvo u reliable 
 
-                        if (user.ImageFile.FileContent.Length > 0) // ako je promenjena slika u reliable upisi je i u blob 
-                        {
-                            CloudBlockBlob blob = await dataRepo.GetBlockBlobReference("users", $"image_{previousUsername}"); // nadji prethodni blok u blobu
-                            await blob.DeleteIfExistsAsync(); // obrisi ga 
+                    if (user.ImageFile != null && user.ImageFile.FileContent != null && user.ImageFile.FileContent.Length > 0) // ako je promenjena slika u reliable upisi je i u blob 
+                    {
+                        CloudBlockBlob blob = await dataRepo.GetBlockBlobReference("users", $"image_{userFromReliable.Id}"); // nadji prethodni blok u blobu
+                        await blob.DeleteIfExistsAsync(); // obrisi ga 
 
-                            CloudBlockBlob newblob = await dataRepo.GetBlockBlobReference("users", $"image_{userFromReliable.Username}"); // kreiraj za ovaj novi username
-                            newblob.Properties.ContentType = userFromReliable.ImageFile.ContentType;
-                            await newblob.UploadFromByteArrayAsync(userFromReliable.ImageFile.FileContent, 0, userFromReliable.ImageFile.FileContent.Length); // upload novu sliku 
-                        }
+                        CloudBlockBlob newblob = await dataRepo.GetBlockBlobReference("users", $"image_{userFromReliable.Id}"); // kreiraj za ovaj novi username
+                        newblob.Properties.ContentType = userFromReliable.ImageFile.ContentType;
+                        await newblob.UploadFromByteArrayAsync(userFromReliable.ImageFile.FileContent, 0, userFromReliable.ImageFile.FileContent.Length); // upload novu sliku 
+                    }
 
-                       await dataRepo.UpdateUser(user, userFromReliable); // sacuva ga u bazu 
-                       await tx.CommitAsync();
-                       return UserMapper.MapUserToFullUserDto(userFromReliable); 
-                    
+                    await dataRepo.UpdateUser(user, userFromReliable); // sacuva ga u bazu 
+                    await tx.CommitAsync();
+                    return UserMapper.MapUserToFullUserDto(userFromReliable);
+
                 }
                 else return new FullUserDTO();
-           
+            }
+
+        }
+
+        public async Task<FullUserDTO> GetUserInfo(Guid id)
+        {
+            var users = await this.StateManager.GetOrAddAsync<IReliableDictionary<Guid, User>>("UserEntities");
+            using (var tx = this.StateManager.CreateTransaction())
+            {
+                ConditionalValue<User> result = await users.TryGetValueAsync(tx, id);
+                if (result.HasValue)
+                {
+                    FullUserDTO user = UserMapper.MapUserToFullUserDto(result.Value);
+                    return user;
+                }
+                else return new FullUserDTO();
 
 
             }
-
         }
     }
 }
